Upgraded to Font Awesome 4.0: Added new icons: circle-arrow-right-alt circle-arrow-left-alt stack-exchange expand-left dot-circle wheelchair vimeo try, turkish-lira (alias) rub, ruble (alias), rouble (alias) - these icons are now using the unicode "f158" which was previously used by the now removed icon "renminbi"
Moved "renminbi" and "cny" to now be alias for "yen", as the icon that was previously used for those codes has been removed. see https://github.com/FortAwesome/Font-Awesome/issues/1338
added the alias "circle-arrow-down-alt" to "download" icon - this is because of the new circle-arrow-right-alt & circle-arrow-left-alt
added the alias "circle-arrow-up-alt" to "upload" icon - see previous explanation
"stackexchange" was assigned to the "stack-overflow" icon, so I changed the code to "stack-overflow" and made the code "stackexchange" to be an alias for the new "stack-exchange" icon
Updated lib.editor to reflect changes.